refactor(client): rename resources/ -> api/
Finding "foo.res" by auto-complete in tools like Ctrl+P in vim was
really annoying as there are so many other internal matches. "foo.api"
stands out way better.

- Rename resources/ -> api/
- Rename foo.res.ts -> foo.api.ts
- Update all references in code.

Closes: d4h/decisions-project#3925
