+++ src/lib/api/attendance.api.ts
@@ -1,5 +1,5 @@
 import { ActivityType, Attendance, AttendanceStatus } from '../models';
-import { DateParameter, Response, Search } from './shared.res';
+import { DateParameter, Response, Search } from './shared.api';
 
 /**
  * GET /team/attendance
