+++ src/lib/api/equipment.api.ts
@@ -1,4 +1,4 @@
-import { DateParameter, Response, Search } from './shared.res';
+import { DateParameter, Response, Search } from './shared.api';
 import { Equipment, ResultStatus } from '../models';
 
 /**
