+++ src/lib/api/index.ts
@@ -1,0 +1,40 @@
+import * as Accounts from './account.api';
+import * as Activities from './activity.api';
+import * as Attendances from './attendance.api';
+import * as Duties from './duty.api';
+import * as Gear from './equipment.api';
+import * as Groups from './group.api';
+import * as InspectionResults from './inspection-result.api';
+import * as Inspections from './inspection.api';
+import * as Members from './member.api';
+import * as Notes from './note.api';
+import * as Photos from './photo.api';
+import * as Repairs from './repair.api';
+import * as Roles from './role.api';
+import * as Teams from './team.api';
+
+export {
+  ApiError,
+  DateParameter,
+  ModelError,
+  Page,
+  Response,
+  Search
+} from './shared.api';
+
+export {
+  Accounts,
+  Activities,
+  Attendances,
+  Duties,
+  Gear,
+  Groups,
+  InspectionResults,
+  Inspections,
+  Members,
+  Notes,
+  Photos,
+  Repairs,
+  Roles,
+  Teams
+};
