+++ src/lib/api/inspection.api.ts
@@ -1,4 +1,4 @@
-import { DateParameter, Response, Search } from './shared.res';
+import { DateParameter, Response, Search } from './shared.api';
 import { Inspection, InspectionInterval, LocationType } from '../models';
 
 /**
