+++ src/lib/api/photo.api.ts
@@ -1,4 +1,4 @@
-import { DateParameter } from './shared.res';
+import { DateParameter } from './shared.api';
 
 /**
  * GET /organisation/image
