--- src/lib/resources/index.ts
@@ -1,40 +1,0 @@
-import * as Accounts from './account.res';
-import * as Activities from './activity.res';
-import * as Attendances from './attendance.res';
-import * as Duties from './duty.res';
-import * as Gear from './equipment.res';
-import * as Groups from './group.res';
-import * as InspectionResults from './inspection-result.res';
-import * as Inspections from './inspection.res';
-import * as Members from './member.res';
-import * as Notes from './note.res';
-import * as Photos from './photo.res';
-import * as Repairs from './repair.res';
-import * as Roles from './role.res';
-import * as Teams from './team.res';
-
-export {
-  ApiError,
-  DateParameter,
-  ModelError,
-  Page,
-  Response,
-  Search
-} from './shared.res';
-
-export {
-  Accounts,
-  Activities,
-  Attendances,
-  Duties,
-  Gear,
-  Groups,
-  InspectionResults,
-  Inspections,
-  Members,
-  Notes,
-  Photos,
-  Repairs,
-  Roles,
-  Teams
-};
