+++ src/lib/services/duty.service.ts
@@ -5,7 +5,7 @@ import { map } from 'rxjs/operators';
 import { API_ROUTES, HttpOptions, RouteConfig } from '../providers';
 import { ApiHttpClient } from '../client/api.client';
 import { ClientModule } from '../client.module';
-import { Duties } from '../resources';
+import { Duties } from '../api';
 import { Duty } from '../models';
 
 @Injectable({ providedIn: ClientModule })
