+++ src/lib/services/note.service.ts
@@ -6,7 +6,7 @@ import { API_ROUTES, HttpOptions, RouteConfig } from '../providers';
 import { ApiHttpClient } from '../client/api.client';
 import { ClientModule } from '../client.module';
 import { Note } from '../models';
-import { Notes } from '../resources';
+import { Notes } from '../api';
 
 @Injectable({ providedIn: ClientModule })
 export class NoteService {
