+++ src/public-api.ts
@@ -1,7 +1,7 @@
+export * from './lib/api';
 export * from './lib/models';
 export * from './lib/permissions';
 export * from './lib/providers';
-export * from './lib/resources';
 export * from './lib/services';
 export * from './lib/tools';
